+++ src/main/java/org/platemate/dto/request/GetTeamMappingRequest.java
@@ -1,9 +1,9 @@
 package org.platemate.dto.request;
 
-public record GetTeamMappingRequest (
+public record GetTeamMappingRequest(
         String nickname,
-        Long latitude,
-        Long longtitude,
+        Float latitude,
+        Float longtitude,
         Long authCode
-){
+) {
 }
